Extract Scraped classes to individual files
This step is required in order for scraper_test to run tests against
MemberPageJp
