+++ lib/member_page_jp.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'scraped'
+
+class MemberPageJp < Scraped::HTML
+  decorator Scraped::Response::Decorator::AbsoluteUrls
+
+  field :image do
+    noko.css('#photo img/@src').text
+  end
+end
