+++ scraper.rb
@@ -7,130 +7,11 @@ require 'pry'
 require 'scraped'
 require 'scraperwiki'
 
-# require 'open-uri/cached'
-# OpenURI::Cache.cache_path = '.cache'
-require 'scraped_page_archive/open-uri'
+require_rel 'lib'
 
-class LetterListPageEn < Scraped::HTML
-  decorator Scraped::Response::Decorator::CleanUrls
-
-  field :members do
-    wanted_rows.map { |tr| fragment tr => MemberRowEn }
-  end
-
-  field :letter_pages do
-    noko.css('p.r + table tr td.c1 a/@href').map(&:text)
-  end
-
-  private
-
-  def all_rows
-    noko.css('h1#TopContents + table tr', 'h1#TopContents + br + table tr')
-  end
-
-  def wanted_rows
-    all_rows.reject { |tr| tr.css('td').empty? }
-  end
-end
-
-class MemberRowEn < Scraped::HTML
-  field :id do
-    File.basename(image, '.jpg')
-  end
-
-  field :name do
-    raw_name.gsub(/M[rs]\./, '').tidy
-  end
-
-  field :gender do
-    return 'male' if raw_name.include?('Mr.')
-    return 'female' if raw_name.include?('Ms.')
-  end
-
-  field :faction do
-    tds[2].children.map(&:text).join(' ').tidy
-  end
-
-  field :image do
-    tds[0].css('img/@src').to_s
-  end
-
-  field :area do
-    tds[3].text.tidy
-  end
-
-  field :term do
-    46
-  end
-
-  field :source do
-    url.to_s
-  end
-
-  private
-
-  def tds
-    noko.css('td')
-  end
-
-  def raw_name
-    tds[1].text
-  end
-end
-
-class LetterListPageJp < Scraped::HTML
-  decorator Scraped::Response::Decorator::CleanUrls
-
-  field :members do
-    noko.xpath('//tr[td[@class="sh1td5"]]').map do |tr|
-      fragment tr => MemberRowJp
-    end
-  end
-
-  field :letter_pages do
-    noko.xpath('//div[@id="breadcrumb"]/following-sibling::p/a/@href').map(&:text) - [url]
-  end
-end
-
-class MemberRowJp < Scraped::HTML
-  field :id do
-    File.basename(source, '.html')
-  end
-
-  field :name__ja do
-    tds[0].text.tidy
-  end
-
-  field :name__ja_hira do
-    tds[1].text.tidy
-  end
-
-  field :faction__ja do
-    tds[2].text.tidy
-  end
-
-  field :area__ja do
-    tds[3].text.tidy
-  end
-
-  field :source do
-    tds[0].css('a/@href').text
-  end
-
-  private
-
-  def tds
-    noko.css('td')
-  end
-end
-
-class MemberPageJp < Scraped::HTML
-  decorator Scraped::Response::Decorator::CleanUrls
-
-  field :image do
-    noko.css('#photo img/@src').text
-  end
-end
+require 'open-uri/cached'
+OpenURI::Cache.cache_path = '.cache'
+# require 'scraped_page_archive/open-uri'
 
 def scrape(h)
   url, klass = h.to_a.first
